fix: 패키지 구조 변경
- app/sys/common으로 구분
