--- src/main/java/com/aluminium/acoe/com/config/SecurityConfig.java
@@ -1,74 +1,0 @@
-package com.aluminium.acoe.com.config;
-
-import com.aluminium.acoe.com.exception.JwtExceptionFilter;
-import com.aluminium.acoe.com.jwt.JwtAccessDeniedHandler;
-import com.aluminium.acoe.com.jwt.JwtAuthenticationEntryPoint;
-import com.aluminium.acoe.com.jwt.JwtSecurityConfig;
-import com.aluminium.acoe.com.jwt.TokenProvider;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.filter.CorsFilter;
-
-@EnableWebSecurity
-@EnableMethodSecurity
-@RequiredArgsConstructor
-@Configuration
-public class SecurityConfig {
-    private final TokenProvider tokenProvider;
-    private final CorsFilter corsFilter;
-    private final JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
-    private final JwtAccessDeniedHandler jwtAccessDeniedHandler;
-    private final JwtExceptionFilter jwtExceptionFilter;
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
-                // Token 방식을 사용하는 방식이기 때문에 CSRF Disable
-                .csrf().disable()
-                .addFilterBefore(corsFilter, UsernamePasswordAuthenticationFilter.class)
-
-                // 401, 403 Exception Handling
-                .exceptionHandling()
-                .authenticationEntryPoint(jwtAuthenticationEntryPoint)
-                .accessDeniedHandler(jwtAccessDeniedHandler)
-
-                .and()
-                .headers()
-                .frameOptions()
-                .sameOrigin()
-
-                // 세션을 사용하지 않기 때문에 STATELESS 설정
-                .and()
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-
-                // 로그인, 회원가입 등 토큰 필요 없이 접근 가능하도록 허용
-                .and()
-                .authorizeHttpRequests()
-                .requestMatchers(
-                        "/api/authenticate",
-                        "/api/signup",
-                        "/v3/api-docs/**",
-                        "/swagger*/**").permitAll()
-                .anyRequest().authenticated()
-
-                .and()
-                .apply(new JwtSecurityConfig(tokenProvider, jwtExceptionFilter));
-
-        return httpSecurity.build();
-    }
-}
