--- src/main/java/com/aluminium/acoe/com/entity/Authority.java
@@ -1,23 +1,0 @@
-package com.aluminium.acoe.com.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "authority")
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class Authority {
-
-    @Id
-    @Column(name = "authority_name", length = 50)
-    private String authorityName;
-}
