--- src/main/java/com/aluminium/acoe/com/entity/User.java
@@ -1,42 +1,0 @@
-package com.aluminium.acoe.com.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.Set;
-
-@Entity
-@Table(name = "`user`")
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-
-    @Id
-    @Column(name = "user_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
-
-    @Column(name = "username", length = 50, unique = true)
-    private String username;
-
-    @Column(name = "password", length = 100)
-    private String password;
-
-    @Column(name = "nickname", length = 50)
-    private String nickname;
-
-    @Column(name = "activated")
-    private boolean activated;
-
-    @ManyToMany
-    @JoinTable(
-            name = "user_authority",
-            joinColumns = {@JoinColumn(name = "user_id", referencedColumnName = "user_id")},
-            inverseJoinColumns = {@JoinColumn(name = "authority_name", referencedColumnName = "authority_name")})
-    private Set<Authority> authorities;
-}
