--- src/main/java/com/aluminium/acoe/com/exception/BusinessInvalidValueException.java
@@ -1,8 +1,0 @@
-package com.aluminium.acoe.com.exception;
-
-public class BusinessInvalidValueException extends RuntimeException {
-
-    public BusinessInvalidValueException(String message) {
-        super(message);
-    }
-}
