--- src/main/java/com/aluminium/acoe/com/exception/NotFoundMemberException.java
@@ -1,19 +1,0 @@
-package com.aluminium.acoe.com.exception;
-
-public class NotFoundMemberException extends RuntimeException {
-    public NotFoundMemberException() {
-        super();
-    }
-
-    public NotFoundMemberException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public NotFoundMemberException(String message) {
-        super(message);
-    }
-
-    public NotFoundMemberException(Throwable cause) {
-        super(cause);
-    }
-}
