--- src/main/java/com/aluminium/acoe/com/jwt/JwtAccessDeniedHandler.java
@@ -1,19 +1,0 @@
-package com.aluminium.acoe.com.jwt;
-
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException {
-        //필요한 권한이 없이 접근하려 할때 403 Forbidden
-        response.sendError(HttpServletResponse.SC_FORBIDDEN);
-    }
-}
