--- src/main/java/com/aluminium/acoe/com/jwt/TokenProvider.java
@@ -1,120 +1,0 @@
-package com.aluminium.acoe.com.jwt;
-
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Date;
-import java.util.stream.Collectors;
-
-/**
- * 토큰의 생성, 유효성 검증
- */
-@Component
-public class TokenProvider implements InitializingBean {
-
-    private final Logger logger = LoggerFactory.getLogger(TokenProvider.class);
-
-    private static final String AUTHORITIES_KEY = "auth";
-
-    private final String secret;
-    private final long tokenValidityInMilliseconds;
-
-    private Key key;
-
-    public TokenProvider(
-            @Value("${jwt.secret}") String secret,
-            @Value("${jwt.token-validity-in-seconds}") long tokenValidityInSeconds) {
-        this.secret = secret;
-        this.tokenValidityInMilliseconds = tokenValidityInSeconds * 1000;
-    }
-
-    /**
-     * Bean이 생성되고 의존성 주입 받은 후에 secret값을 Base64 Decode해서 key변수에 할당
-     */
-    @Override
-    public void afterPropertiesSet() {
-        byte[] keyBytes = Decoders.BASE64.decode(secret);
-        this.key = Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    /**
-     * Authentication 객체의 권한 정보를 이용해 토큰을 생성하는 함수
-     */
-    public String createToken(Authentication authentication) {
-        String authorities = authentication.getAuthorities().stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.joining(","));
-
-        // 설정한 토큰 만료 시간 설정
-        long now = (new Date()).getTime();
-        Date validity = new Date(now + this.tokenValidityInMilliseconds);
-
-        // 토큰 반환
-        return Jwts.builder()
-                .setSubject(authentication.getName())
-                .claim(AUTHORITIES_KEY, authorities)
-                .signWith(key, SignatureAlgorithm.HS512)
-                .setExpiration(validity)
-                .compact();
-    }
-
-    /**
-     * 토큰에 담긴 정보를 이용해 Authentication 객체를 반환하는 함수
-     */
-    public Authentication getAuthentication(String token) {
-        // 토큰으로 클레임을 만듬
-        Claims claims = Jwts
-                .parserBuilder()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-
-        Collection<? extends GrantedAuthority> authorities =
-                Arrays.stream(claims.get(AUTHORITIES_KEY).toString().split(","))
-                        .map(SimpleGrantedAuthority::new)
-                        .collect(Collectors.toList());
-
-        // User객체를 만듦
-        User principal = new User(claims.getSubject(), "", authorities);
-
-        // Authentication 객체 반환
-        return new UsernamePasswordAuthenticationToken(principal, token, authorities);
-    }
-
-    /**
-     * 토큰 유효성 검증 (정상 true)
-     */
-//    public boolean validateToken(String token) {
-//        try {
-//            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-//            return true;
-//        } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
-//            logger.info("잘못된 JWT 서명입니다.");
-//        } catch (ExpiredJwtException e) {
-//            logger.info("만료된 JWT 토큰입니다.");
-//        } catch (UnsupportedJwtException e) {
-//            logger.info("지원되지 않는 JWT 토큰입니다.");
-//        } catch (IllegalArgumentException e) {
-//            logger.info("JWT 토큰이 잘못되었습니다.");
-//        }
-//        return false;
-//    }
-    public void validateToken(String token) {
-        Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-    }
-}
