--- src/main/java/com/aluminium/acoe/com/repository/AuthorityRepository.java
@@ -1,9 +1,0 @@
-package com.aluminium.acoe.com.repository;
-
-import com.aluminium.acoe.com.entity.Authority;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AuthorityRepository extends JpaRepository<Authority, String> {
-}
