--- src/main/java/com/aluminium/acoe/com/service/CustomUserDetailsService.java
@@ -1,53 +1,0 @@
-package com.aluminium.acoe.com.service;
-
-import com.aluminium.acoe.com.repository.UserRepository;
-import com.aluminium.acoe.com.entity.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Component("userDetailsService")
-public class CustomUserDetailsService implements UserDetailsService {
-    private final UserRepository userRepository;
-
-    public CustomUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    /**
-     * 로그인 -> 유저, 권한 정보 반환
-     * @param username the username identifying the user whose data is required.
-     * @return
-     */
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(final String username) {
-        return userRepository.findOneWithAuthoritiesByUsername(username)
-                .map(user -> createUser(username, user))
-                .orElseThrow(() -> new UsernameNotFoundException(username + " -> 데이터베이스에서 찾을 수 없습니다."));
-    }
-
-    /**
-     * User 객체 생성 후 반환
-     * @param username
-     * @param user
-     * @return
-     */
-    private org.springframework.security.core.userdetails.User createUser(String username, User user) {
-        if (!user.isActivated())
-            throw new RuntimeException(username + " -> 활성화되어 있지 않습니다.");
-
-        List<GrantedAuthority> grantedAuthorities = user.getAuthorities().stream()
-                .map(authority -> new SimpleGrantedAuthority(authority.getAuthorityName()))
-                .collect(Collectors.toList());
-
-        return new org.springframework.security.core.userdetails.User(user.getUsername(), user.getPassword(), grantedAuthorities);
-    }
-}
