--- src/main/java/com/aluminium/acoe/com/util/SecurityUtil.java
@@ -1,39 +1,0 @@
-package com.aluminium.acoe.com.util;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Optional;
-
-public class SecurityUtil {
-
-    private static final Logger logger = LoggerFactory.getLogger(SecurityUtil.class);
-
-    private SecurityUtil() {
-    }
-
-    // Security Context의 Authentication 객체를 활용해 username을 반환
-    public static Optional<String> getCurrentUsername() {
-        // SecurityContext에서 Authentication 객체 가져옴
-        final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null) {
-            logger.debug("Security Context에 인증 정보가 없습니다.");
-            return Optional.empty();
-        }
-
-        // username 가져오기
-        String username = null;
-        if (authentication.getPrincipal() instanceof UserDetails) {
-            UserDetails springSecurityUser = (UserDetails) authentication.getPrincipal();
-            username = springSecurityUser.getUsername();
-        } else if (authentication.getPrincipal() instanceof String) {
-            username = (String) authentication.getPrincipal();
-        }
-
-        return Optional.ofNullable(username);
-    }
-}
