+++ src/main/java/com/aluminium/acoe/common/exception/BusinessInvalidValueException.java
@@ -1,0 +1,8 @@
+package com.aluminium.acoe.common.exception;
+
+public class BusinessInvalidValueException extends RuntimeException {
+
+    public BusinessInvalidValueException(String message) {
+        super(message);
+    }
+}
