+++ src/main/java/com/aluminium/acoe/common/exception/DuplicateMemberException.java
@@ -1,0 +1,19 @@
+package com.aluminium.acoe.common.exception;
+
+public class DuplicateMemberException extends RuntimeException {
+    public DuplicateMemberException() {
+        super();
+    }
+
+    public DuplicateMemberException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DuplicateMemberException(String message) {
+        super(message);
+    }
+
+    public DuplicateMemberException(Throwable cause) {
+        super(cause);
+    }
+}
