+++ src/main/java/com/aluminium/acoe/common/exception/GlobalExceptionHandler.java
@@ -1,0 +1,181 @@
+package com.aluminium.acoe.common.exception;
+
+import com.aluminium.acoe.common.api.code.ErrorCode;
+import com.aluminium.acoe.common.api.response.ErrorResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.HttpRequestMethodNotSupportedException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.MissingRequestHeaderException;
+import org.springframework.web.bind.MissingServletRequestParameterException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.client.HttpClientErrorException;
+import org.springframework.web.servlet.NoHandlerFoundException;
+
+/**
+ * Controller 내에서 발생하는 Exception 대해서 Catch 하여 응답값(Response)을 보내주는 기능을 수행함.
+ */
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    private final HttpStatus HTTP_STATUS_OK = HttpStatus.OK;
+
+    /**
+     * [Exception] API 호출 시 '객체' 혹은 '파라미터' 데이터 값이 유효하지 않은 경우
+     *
+     * @param ex MethodArgumentNotValidException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    protected ResponseEntity<ErrorResponse> handleMethodArgumentNotValidException(MethodArgumentNotValidException ex) {
+        log.error("handleMethodArgumentNotValidException", ex);
+        BindingResult bindingResult = ex.getBindingResult();
+        StringBuilder stringBuilder = new StringBuilder();
+        for (FieldError fieldError : bindingResult.getFieldErrors()) {
+            stringBuilder.append(fieldError.getField()).append(":");
+            stringBuilder.append(fieldError.getDefaultMessage());
+            stringBuilder.append(", ");
+        }
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.NOT_VALID_ERROR, String.valueOf(stringBuilder));
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+    /**
+     * [Exception] API 호출 시 지원하지 않은 HTTP method 호출 할 경우 발생
+     *
+     * @param ex MissingRequestHeaderException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(HttpRequestMethodNotSupportedException.class)
+    protected ResponseEntity<ErrorResponse> handleHttpRequestMethodNotSupportedException(HttpRequestMethodNotSupportedException ex) {
+        log.error("handleHttpRequestMethodNotSupportedException", ex);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.NOT_ALLOWED_METHOD, ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.METHOD_NOT_ALLOWED);
+    }
+
+    /**
+     * [Exception] API 호출 시 'Header' 내에 데이터 값이 유효하지 않은 경우
+     *
+     * @param ex MissingRequestHeaderException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(MissingRequestHeaderException.class)
+    protected ResponseEntity<ErrorResponse> handleMissingRequestHeaderException(MissingRequestHeaderException ex) {
+        log.error("handleMissingRequestHeaderException", ex);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.REQUEST_HEADER_MISSING_ERROR, ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+    /**
+     * [Exception] 클라이언트에서 Body로 '객체' 데이터가 넘어오지 않았을 경우
+     *
+     * @param ex HttpMessageNotReadableException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(HttpMessageNotReadableException.class)
+    protected ResponseEntity<ErrorResponse> handleHttpMessageNotReadableException(
+        HttpMessageNotReadableException ex) {
+        log.error("HttpMessageNotReadableException", ex);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.REQUEST_BODY_MISSING_ERROR, ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+    /**
+     * [Exception] 클라이언트에서 request로 '파라미터로' 데이터가 넘어오지 않았을 경우
+     *
+     * @param ex MissingServletRequestParameterException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(MissingServletRequestParameterException.class)
+    protected ResponseEntity<ErrorResponse> handleMissingRequestHeaderExceptionException(
+        MissingServletRequestParameterException ex) {
+        log.error("handleMissingServletRequestParameterException", ex);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.MISSING_REQUEST_PARAMETER_ERROR, ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+
+    /**
+     * [Exception] 잘못된 서버 요청일 경우 발생한 경우
+     *
+     * @param e HttpClientErrorException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(HttpClientErrorException.BadRequest.class)
+    protected ResponseEntity<ErrorResponse> handleBadRequestException(HttpClientErrorException e) {
+        log.error("HttpClientErrorException.BadRequest", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.BAD_REQUEST_ERROR, e.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+
+    /**
+     * [Exception] 잘못된 주소로 요청 한 경우
+     *
+     * @param e NoHandlerFoundException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(NoHandlerFoundException.class)
+    protected ResponseEntity<ErrorResponse> handleNoHandlerFoundExceptionException(NoHandlerFoundException e) {
+        log.error("handleNoHandlerFoundExceptionException", e);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.NOT_FOUND_ERROR, e.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+    }
+
+
+    // =========  Business Exception ===============
+
+    @ExceptionHandler({BusinessInvalidValueException.class})
+    protected ResponseEntity<ErrorCode> handleBusinessInvalidValueException(BusinessInvalidValueException e) {
+        log.error("handleBusinessInvalidValueException", e);
+        ErrorResponse response = ErrorResponse.of(ErrorCode.BUISINESS_INVALID_VALUE, e.getMessage());
+        return new ResponseEntity(response, HttpStatus.BAD_REQUEST);
+    }
+
+    /**
+     * [Exception] 로그인 not found member  Excetpion
+     *
+     * @param ex NotFoundMemberException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(NotFoundMemberException.class)
+    protected ResponseEntity<ErrorResponse> handleNotFoundMemberException(NotFoundMemberException ex) {
+        log.error("handleNotFoundMemberException", ex);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.NOT_FOUND_MEMBER, ex.getMessage());
+        return new ResponseEntity<>(response, HTTP_STATUS_OK);
+    }
+
+    /**
+     * [Exception] 로그인 duplicate member  Excetpion
+     *
+     * @param ex NotFoundMemberException
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(DuplicateMemberException.class)
+    protected ResponseEntity<ErrorResponse> handleDuplicateMemberException(DuplicateMemberException ex) {
+        log.error("handleDuplicateMemberException", ex);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.DUPLICATE_MEMBER, ex.getMessage());
+        return new ResponseEntity<>(response, HTTP_STATUS_OK);
+    }
+
+    // ============= 최상위 Exception ===================
+
+    /**
+     * [Exception] 모든 Exception 경우 발생
+     *
+     * @param ex Exception
+     * @return ResponseEntity<ErrorResponse>
+     */
+    @ExceptionHandler(Exception.class)
+    protected final ResponseEntity<ErrorResponse> handleAllExceptions(Exception ex) {
+        log.error("Exception", ex);
+        final ErrorResponse response = ErrorResponse.of(ErrorCode.INTERNAL_SERVER_ERROR, ex.getMessage());
+        return new ResponseEntity<>(response, HTTP_STATUS_OK);
+    }
+}
