+++ src/main/java/com/aluminium/acoe/common/exception/NotFoundMemberException.java
@@ -1,0 +1,19 @@
+package com.aluminium.acoe.common.exception;
+
+public class NotFoundMemberException extends RuntimeException {
+    public NotFoundMemberException() {
+        super();
+    }
+
+    public NotFoundMemberException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotFoundMemberException(String message) {
+        super(message);
+    }
+
+    public NotFoundMemberException(Throwable cause) {
+        super(cause);
+    }
+}
