+++ src/main/java/com/aluminium/acoe/sys/controller/AuthController.java
@@ -1,0 +1,65 @@
+package com.aluminium.acoe.sys.controller;
+
+import com.aluminium.acoe.sys.dto.LoginDto;
+import com.aluminium.acoe.sys.dto.TokenDto;
+import com.aluminium.acoe.sys.jwt.JwtFilter;
+import com.aluminium.acoe.sys.jwt.TokenProvider;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import jakarta.validation.Valid;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api")
+public class AuthController {
+
+    private final TokenProvider tokenProvider;
+    private final AuthenticationManagerBuilder authenticationManagerBuilder;
+
+    public AuthController(TokenProvider tokenProvider, AuthenticationManagerBuilder authenticationManagerBuilder) {
+        this.tokenProvider = tokenProvider;
+        this.authenticationManagerBuilder = authenticationManagerBuilder;
+    }
+
+    /**
+     * 로그인 API
+     */
+    @PostMapping("/authenticate")
+    @Operation(summary = "로그인", description  = "Username, Password 인증 후 토큰 반환", responses = {
+            @ApiResponse(responseCode = "200", description = "로그인 성공", content = @Content(schema = @Schema(implementation = TokenDto.class)))
+    })
+    @Parameter(name = "loginDto", description = "접속 계정 정보 객체", in = ParameterIn.DEFAULT)
+    public ResponseEntity<TokenDto> authorize(@Valid @RequestBody LoginDto loginDto) {
+        // ID, PW를 받아 토큰 생성
+        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(loginDto.getUsername(), loginDto.getPassword());
+
+        // 토큰을 이용해 Authentication 객체 생성 -> loadUserByUsername 실행
+        Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
+        // 객체를 SecurityContext에 저장
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+
+        // 객체를 createToken으로 JWT 토큰 생성
+        String jwt = tokenProvider.createToken(authentication);
+
+        // JWT 토큰을 Response Header에 넣음
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.add(JwtFilter.AUTHORIZATION_HEADER, "Bearer " + jwt);
+
+        // JWT 토큰을 TokenDto를 이용해 Response Body에도  넣음
+        return new ResponseEntity<>(new TokenDto(jwt), httpHeaders, HttpStatus.OK);
+    }
+}
