+++ src/main/java/com/aluminium/acoe/sys/jwt/JwtAuthenticationEntryPoint.java
@@ -1,0 +1,20 @@
+package com.aluminium.acoe.sys.jwt;
+
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Component
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+    @Override
+    public void commence(HttpServletRequest request,
+                         HttpServletResponse response,
+                         AuthenticationException authException) throws IOException {
+        // 유효한 자격증명을 제공하지 않고 접근하려 할때 401 Unauthorized Error
+        response.sendError(HttpServletResponse.SC_UNAUTHORIZED);
+    }
+}
