+++ src/main/java/com/aluminium/acoe/sys/jwt/JwtFilter.java
@@ -1,0 +1,60 @@
+package com.aluminium.acoe.sys.jwt;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.util.StringUtils;
+import org.springframework.web.filter.GenericFilterBean;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+
+import java.io.IOException;
+
+public class JwtFilter extends GenericFilterBean {
+
+    private static final Logger logger = LoggerFactory.getLogger(JwtFilter.class);
+    public static final String AUTHORIZATION_HEADER = "Authorization";
+    private TokenProvider tokenProvider;
+
+    public JwtFilter(TokenProvider tokenProvider) {
+        this.tokenProvider = tokenProvider;
+    }
+
+    /**
+     * 토큰의 인증 정보를 SecurityContext에 저장하는 역할
+     */
+    @Override
+    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+        HttpServletRequest httpServletRequest = (HttpServletRequest) servletRequest;
+        // Request에서 토큰 가져옴
+        String jwt = resolveToken(httpServletRequest);
+        String requestURI = httpServletRequest.getRequestURI();
+
+        // jwt 토큰 유효성 검증 후 정상이면 Security Context에 저장
+        if (StringUtils.hasText(jwt)) {
+            tokenProvider.validateToken(jwt);
+
+            Authentication authentication = tokenProvider.getAuthentication(jwt);
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+            logger.debug("Security Context에 '{}' 인증 정보를 저장했습니다, uri: {}", authentication.getName(), requestURI);
+        } else
+            logger.debug("유효한 JWT 토큰이 없습니다, uri: {}", requestURI);
+
+        filterChain.doFilter(servletRequest, servletResponse);
+    }
+
+    /**
+     * Request Header에서 토큰 정보를 가져오는 함수
+     */
+    private String resolveToken(HttpServletRequest request) {
+        String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
+        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith("Bearer "))
+            return bearerToken.substring(7);
+        return null;
+    }
+}
