+++ src/main/java/com/aluminium/acoe/sys/repository/AuthorityRepository.java
@@ -1,0 +1,9 @@
+package com.aluminium.acoe.sys.repository;
+
+import com.aluminium.acoe.sys.entity.Authority;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AuthorityRepository extends JpaRepository<Authority, String> {
+}
